Add SoftButton common component
