+++ src/components/common/SoftButton.tsx
@@ -1,0 +1,13 @@
+import styled from "styled-components/native"
+
+export const SoftButtonView = styled.TouchableOpacity`
+  justify-content: center;
+  align-items: center;
+  padding: 14px 35px;
+  border-radius: 14px;
+`
+
+export const SoftButtonText = styled.Text`
+  font-size: 18px;
+  font-weight: 800;
+`
