+++ src/components/common/index.ts
@@ -1,6 +1,14 @@
 import Page from "./Page"
+import { SoftButtonView, SoftButtonText } from "./SoftButton"
 import BackButton from "./BackButton"
 import CloseButton from "./CloseButton"
 import LogoIcon from "./LogoIcon"
 
-export { Page, BackButton, CloseButton, LogoIcon }
+export {
+  Page,
+  BackButton,
+  CloseButton,
+  LogoIcon,
+  SoftButtonView,
+  SoftButtonText,
+}
